Extract AbstractPrestoResultSet
`AbstractPrestoResultSet` implements most of the `ResultSet` interface,
while leaving server interactions to the `PrestoResultSet` class.
