+++ presto-jdbc/src/main/java/io/prestosql/jdbc/PrestoPreparedStatement.java
@@ -56,6 +56,9 @@ import java.util.regex.Pattern;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.io.BaseEncoding.base16;
 import static io.prestosql.client.ClientTypeSignature.VARCHAR_UNBOUNDED_LENGTH;
+import static io.prestosql.jdbc.AbstractPrestoResultSet.DATE_FORMATTER;
+import static io.prestosql.jdbc.AbstractPrestoResultSet.TIMESTAMP_FORMATTER;
+import static io.prestosql.jdbc.AbstractPrestoResultSet.TIME_FORMATTER;
 import static io.prestosql.jdbc.ColumnInfo.setTypeInfo;
 import static io.prestosql.jdbc.ObjectCasts.castToBigDecimal;
 import static io.prestosql.jdbc.ObjectCasts.castToBinary;
@@ -69,9 +72,6 @@ import static io.prestosql.jdbc.ObjectCasts.castToLong;
 import static io.prestosql.jdbc.ObjectCasts.castToShort;
 import static io.prestosql.jdbc.ObjectCasts.castToTime;
 import static io.prestosql.jdbc.ObjectCasts.castToTimestamp;
-import static io.prestosql.jdbc.PrestoResultSet.DATE_FORMATTER;
-import static io.prestosql.jdbc.PrestoResultSet.TIMESTAMP_FORMATTER;
-import static io.prestosql.jdbc.PrestoResultSet.TIME_FORMATTER;
 import static java.lang.Long.parseLong;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
